Extract drag behaviour into a separate object
