+++ lib/overlap.detector.js
